refactor(summary-notification): rename notification_interface.py to notifier.py
インターフェース命名規則の整理（-er/-or 名詞ルール）。
クラス名 INotifier は変更なし。
