+++ lambda/summary-notification/src/domain/notifier.py
